# lib/api-zod/src/generated/types/kycTaxFormInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 7of1 API specification
 * OpenAPI spec version: 0.1.0
 */
import type { KycTaxFormType } from './kycTaxFormType';

export interface KycTaxFormInput {
  taxFormType: KycTaxFormType;
  /** Supabase Storage path returned by /kyc/upload-url */
  storagePath: string;
}